Fix docker caching
Oh, it’s because actions was using # syntax=docker/dockerfile:1.3 syntax which allows variable expansion in RUN --mount, and the mount had the wrong permissions, and locally it was using # syntax=docker/dockerfile:1.2 which does not. So the previous ‘fix’ really just broke the conda package caching because it was then trying to cache ${CONDA_DIR}/pkgs rather than interpreting that into /srv/conda/pkgs. That was a valid path locally, but not accessed by conda, so no explosions.